checagem de erros na musica

+++ src/domains/Musica/services/MusicaService.ts
@@ -1,5 +1,7 @@
 import { Musica } from "@prisma/client";
 import prisma from "../../../../config/prismaClient";
+import ArtistaService from "../../Artista/services/ArtistaService";
+import { QueryError } from "../../../../errors/QueryError";
 
 class MusicaService {
     async create(body: Musica){
@@ -18,13 +20,17 @@ class MusicaService {
     }
 
     async findFromArtist(id: number){
-        return await prisma.musica.findMany({
+        const musicasArtista = await prisma.musica.findMany({
             where: { ArtistaID: id},
             include: {
                 Artista: true,
                 Album: true,
             }
         })
+        if(!musicasArtista){
+            throw new QueryError('Esse artista não existe/não tem musicas')
+        }
+        return musicasArtista
     }
 
     async findAll() {
@@ -37,16 +43,29 @@ class MusicaService {
     }
 
     async findById(id: number) {
-        return await prisma.musica.findUnique({
+        const existingMusica = await prisma.musica.findUnique({
+            where: {ID_Musica:id}
+        })
+        if(!existingMusica){
+            throw new QueryError('Não existe musica com esse id')
+        }
+        const musica = await prisma.musica.findUnique({
             where: { ID_Musica: id },
             include: {
                 Artista: true,
                 Album: true,
             },
         });
+        return musica
     }
 
     async update(id: number, body: Partial<Musica>) {
+        const existingMusica = await prisma.musica.findUnique({
+            where: {ID_Musica:id}
+        })
+        if(!existingMusica){
+            throw new QueryError('Não existe musica com esse id')
+        }
         const updatedMusica = await prisma.musica.update({
             where: { ID_Musica: id },
             data: body,
@@ -55,6 +74,12 @@ class MusicaService {
     }
 
     async delete(id: number) {
+        const existingMusica = await prisma.musica.findUnique({
+            where: {ID_Musica:id}
+        })
+        if(!existingMusica){
+            throw new QueryError('Não existe musica com esse id')
+        }
         const deletedMusica = await prisma.musica.delete({
             where: { ID_Musica: id },
         });
